Continued general refactoring of code. Added more code for Nuke
Now the NUKE_PATH variable is set in "project_env.sh" so we
actually have a Nuke environment!
