+++ .gitignore
@@ -1,5 +1,6 @@
 *~
 *.swp
+*.swo
 *.pyc
 *.pyc
 *.mb
