+++ nuke-tools/init.py
@@ -1,0 +1,6 @@
+import nuke
+
+# Add subfolders to path to make this a little cleaner
+nuke.pluginAddPath( './gizmos' )
+nuke.pluginAddPath( './python' )
+nuke.pluginAddPath( './plugins' )
