+++ nuke-tools/menu.py
@@ -1,0 +1,2 @@
+# This file should only contain code that will be run during interactive sessions
+# It will not be run during terminal or render session.
