+++ nuke-tools/plugins/init.py
@@ -1,0 +1,3 @@
+import sys
+
+sys.stderr.write( "\nCustom Owned Nuke plugins loaded...\n" )
