+++ nuke-tools/python/init.py
@@ -1,0 +1,3 @@
+import sys
+
+sys.stderr.write( "\nCustom Owned Nuke scripts loaded...\n" )
