+++ project_env.sh
@@ -133,5 +133,15 @@ export MAYA_SHELF_DIR=${MAYA_TOOLS_DIR}/shelf
 export MAYA_SCRIPT_PATH=${MAYA_SCRIPT_PATH}:${PYTHONPATH}:${MAYA_SHELF_DIR}
 
 ###############################################################################
+# Nuke specific environment
+###############################################################################
+
+export NUKE_LOCATION=/usr/local/Nuke6.3v9
+
+export NUKE_TOOLS_DIR=${PROJECT_TOOLS}/nuke-tools
+export NUKE_PATH=${HOME}/.nuke:${NUKE_TOOLS_DIR}:${NUKE_LOCATION}/plugins/user:${NUKE_LOCATION}/plugins/icons:${NUKE_LOCATION}/plugins
+
+###############################################################################
 # BEGIN AWESOMENESS!!!
 ###############################################################################
+
