+++ python2.6libs/hou_asset_mgr.py
@@ -6,7 +6,6 @@ import sqlite3 as lite
 import os, glob
 import hou
 import subprocess
-import tempfile
 from ui_tools import ui, messageSeverity, fileMode
 
 import utilities as amu #asset manager utilites
@@ -188,27 +187,6 @@ def lockAsset(node, lockit):
         opts.setLockContents(lockit)
         ndef.setOptions(opts)
 
-def lockedBy(logname):
-    """Returns the true name of based on the login name passed in."""
-    tfd = tempfile.NamedTemporaryFile(mode='r+') # Create temp file to write to
-    myargs = ["/usr/bin/ldapsearch", "-LLL", "uid=" + str(logname), "cn"] # Command args
-    try:
-        subprocess.check_call(myargs, executable="ldapsearch", stdout=tfd)
-        tfd.seek(0) # Return to start of file
-        fstr = str(tfd.read()) # Read contents of file and cast to string
-        tfd.close() # Close and delete our temp file
-        fstr = fstr.strip() # Strip leading and trailing whitespace
-        lastline = fstr.splitlines()[-1] # Get last line
-        truename = lastline[4:] # Strip off first four characters of line
-
-        lockstr = "This asset is locked by the following user:\n\n"
-        lockstr += "User Name: " + logname + "\n"
-        lockstr += "Real Name: " + truename
-        return lockstr # Return lock string
-    except Exception as ex:
-        exstr = "Encountered exception: " + str(ex) + "\nUser's name not found."# Uh-oh... We jacked something up...
-        return exstr
-
 def get_filename(parentdir):
     return os.path.basename(os.path.dirname(parentdir))+'_'+os.path.basename(parentdir)
 
@@ -275,7 +253,10 @@ def checkout(node = None):
                 lockOTL(filename)
                 ui.infoWindow("Checkout Successful!", wtitle='Success!')
             else:
-                ui.infoWindow(lockedBy(info[3].encode('utf-8')))
+                logname, realname = amu.lockedBy(info[3].encode('utf-8'))
+                whoLocked = 'User Name: ' + logname + '\nReal Name: ' + realname + '\n'
+                errstr = 'Cannot checkout asset. Locked by: \n\n' + whoLocked
+                ui.infoWindow(errstr, wtitle='Asset Locked', msev=messageSeverity.Error)
     else:
         #ui.infoWindow("Select EXACTLY one node.")
         checkoutLightingFile()
@@ -302,7 +283,10 @@ def checkin(node = None):
                     unlockOTL(filename)
                     ui.infoWindow("Checkin Successful!")
                 else:
-                    ui.infoWindow(lockedBy(info[3].encode('utf-8')))
+                    logname, realname = amu.lockedBy(info[3].encode('utf-8'))
+                    whoLocked = 'User Name: ' + logname + '\nReal Name: ' + realname + '\n'
+                    errstr = 'Cannot checkin asset. Locked by: \n\n' + whoLocked
+                    ui.infoWindow(errstr, wtitle='Asset Locked', msev=messageSeverity.Error)
             else:
                 ui.infoWindow("Already checked in.")
     else:
@@ -428,7 +412,10 @@ def rename(node = None):
                                 os.system('rm -f '+oldlibraryPath)
                                 amu.renameAsset(assetDirPath, newfilename)
                 else:
-                    ui.infoWindow(lockedBy(info[3].encode('utf-8')))
+                    logname, realname = amu.lockedBy(info[3].encode('utf-8'))
+                    whoLocked = 'User Name: ' + logname + '\nReal Name: ' + realname + '\n'
+                    errstr = 'Cannot checkout asset. Locked by: \n\n' + whoLocked
+                    ui.infoWindow(errstr, wtitle='Asset Locked', msev=messageSeverity.Error)
     else:
         ui.infoWindow("Select EXACTLY one node.")
 
@@ -456,7 +443,10 @@ def deleteAsset(node = None):
 
                 info = getFileInfo(oldfilename)
                 if info[2]:
-                    ui.infoWindow(lockedBy(info[3].encode('utf-8')), wtitle='Asset Locked', msev=messageSeverity.Error)
+                    logname, realname = amu.lockedBy(info[3].encode('utf-8'))
+                    whoLocked = 'User Name: ' + logname + '\nReal Name: ' + realname + '\n'
+                    errstr = 'Cannot delete asset. Locked by: \n\n' + whoLocked
+                    ui.infoWindow(errstr, wtitle='Asset Locked', msev=messageSeverity.Error)
                     return
 
                 if not amu.canRemove(assetDirPath):
